refactor: pull target structs into their own files
