+++ src/commands/publish/upload_form/mod.rs
@@ -11,7 +11,6 @@ use crate::commands::build::wranglerjs;
 use crate::commands::kv::bucket::AssetManifest;
 use crate::settings::binding;
 use crate::settings::metadata::Metadata;
-use crate::settings::target::kv_namespace;
 use crate::settings::target::{Target, TargetType};
 
 use project_assets::ProjectAssets;
